refactor: improve app structure to make code more readable
As layout is not really a component and also nothing reusable and has
quite some logic to it, i have decided to move out of the components
folder and now work as a standalone file and a bridge between the app
file and the rest of the app structure

As I would say max width is also not a component and its only purpose is
to set some max width on the content, I have decided to create a
max-width HOC and wrap the layout in it.

also moving the header in the app folder as it now makes sence to use it
there
