--- frontend/src/components/max-width/MaxWidth.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import { Section, Wrapper } from './styles';
-
-const MaxWidth = ({ children }) => (
-  <>
-    <Section>
-      <Wrapper>
-        {children}
-      </Wrapper>
-    </Section>
-  </>
-);
-
-export default MaxWidth;
