+++ frontend/src/hoc/withMaxWidth/styles.js
