+++ frontend/src/hoc/withMaxWidth/withMaxWidth.js
@@ -1,0 +1,20 @@
+import React from 'react';
+import { Section, Wrapper } from './styles';
+
+const WithMaxWidth = (Component) => {
+
+  // eslint-disable-next-line react/display-name
+  return () => {
+    return (
+      <>
+        <Section>
+          <Wrapper>
+            <Component />
+          </Wrapper>
+        </Section>
+      </>    
+    )
+  }
+}
+
+export default WithMaxWidth;
